Исправил основные недочеты. Оставил без изменений соответствие типа вагона и поезда, ожидая дополнительного комментария. Мне кажется, что вынос типа вагона в родительский класс не решит проблему переноса метода т.к вагоны не аккумулируются в родительском классе, поскольку создаются через New. В таком случае будет затруднительно оперировать ими как единым массивом
